Update README.md
The movie explores themes like dreams vs. reality, grief, and the power of the subconscious mind. It’s known for its innovative storytelling and non-linear narrative, which keeps viewers engaged and guessing. The iconic "Paris folding scene" showcases groundbreaking visual effects, making it a hallmark of modern cinema. Inception is often regarded as one of Christopher Nolan’s masterpieces and has a lasting legacy in the sci-fi genre.